refactor sys directory structure and added iterators
These changes form part of a WIP to introduce cross-platform event
queues.

A large amount of changes were introduced in this commit.

Previously, we had a situation whereby sys::Events was wrapped by
interfaces::Event and this type was then being used in ffi syscalls.

This created a weird circular relationship between the sys and
interfaces modules. It was also hard to understand if an `interfaces` of
`sys` type was being used as the name `Event` and `Events` was being
used by types defined in both of these modules.

With above in mind, `sys` types were renamed and only `sys` Event types
are used in the ffi syscalls.

Iterators were also introduced and will be required then a readiness
list is being returned from either epoll or kqueue. The iterators will
yield `interfaces` types, which proxy to OS specific implementations.