archived some of the kqueue based examples
These will need to be reworked somehow (if possible). The latest changes
do not expose the `ffi` module directly, but makes use of the `Interest`
type instead. I don't believe the timer one will be possible to use as
an event source. `mio` fundamentally is a reactor specific for I/O
operations for reading and writing to sockets and pipes. So might need a
different reactor / design for the timer.